Improve bound and unbound signal hints
`pyqtSignal` is a descriptor and does not have `emit`, `connect`, and `disconnect` methods.  Rather it is used as a class attribute and when you access `an_instance.the_signal` you get a `pyqtBoundSignal` instead of the original `pyqtSignal`.  So the bound signal needs those methods and the unbound signal needs `__get__()`.